Subject: [PATCH] fix: MSVC C4267 warning-as-error in socks5-client.c
On Windows, send() expects int for the third argument (len), but
socks5-client.c was casting the ssize_t value to (size_t).  MSVC
at /W3 warns C4267 about the implicit narrowing from size_t (64-bit)
to int (32-bit), and with /WX the warning becomes a fatal error
(C2220).

Fix the two send() sites by adding an explicit (int) cast guarded
with #if defined(WIN32), matching the existing pattern used in
output.c.

AG: use helper and fixup similar cases

Ref: txiki.js CI build (error C2220 at socks5-client.c:239,64)

+++ include/libwebsockets.h
@@ -212,7 +212,7 @@ typedef pthread_mutex_t lws_mutex_t;
 #endif
 #endif /* freertos */
 
-#define LWS_POSIX_LENGTH_CAST(x) (x)
+#define LWS_POSIX_LENGTH_CAST(x) ((size_t)(x))
 
 #if defined(LWS_HAVE_SYS_CAPABILITY_H) && defined(LWS_HAVE_LIBCAP)
 #include <sys/capability.h>
